// types.ts
// types.ts
export interface FoodItem {
  id?: string; // Optional for input, as id is generated by Firestore
  name: string;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  portion?: string;
  timestamp: number; // milliseconds since epoch
}

export interface MacroGoals {
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
}

export interface UserProfile {
  displayName?: string | null;
  photoURL?: string | null;
  email?: string | null;
  createdAt?: number;
  lastSeen?: number;
  macroGoals?: MacroGoals;
}